backend/requirements.txt
Django==4.2.0
djangorestframework==3.14.0
django-cors-headers==4.1.0
django-ckeditor==6.5.1
python-dotenv==1.0.0
deep-translator==1.11.4
httpx==0.13.3
django-redis==5.2.0
python-decouple==3.8
